refactor lift subsystem to elevator and dispenser; update related commands and imports
desync issues
