+++ src/main/java/frc/robot/commands/intake/Dislodge.java
@@ -3,28 +3,29 @@ package frc.robot.commands.intake;
 import frc.robot.commands.util.FunctionWrapper;
 import frc.robot.constants.DispenserConstants;
 import frc.robot.constants.IntakeConstants;
+import frc.robot.subsystems.dispenser.Dispenser;
 import frc.robot.subsystems.intake.Intake;
-import frc.robot.subsystems.lift.Lift;
 
 public class Dislodge extends FunctionWrapper {
     /**
      * Reverse until the beam brake is freed
      *
-     * @param intake the intake subsystem
+     * @param intake    the intake subsystem
+     * @param dispenser the dispenser subsystem
      */
-    public Dislodge(Intake intake, Lift lift) {
+    public Dislodge(Intake intake, Dispenser dispenser) {
         super(
             () -> {
                 intake.set(IntakeConstants.Power.kDislodge);
-                lift.dispense(DispenserConstants.Power.kDislodge);
+                dispenser.dispense(DispenserConstants.Power.kDislodge);
             },
             () -> !intake.getRoller().broken,
             () -> {
                 intake.stop();
-                lift.dispense(0);
+                dispenser.dispense(0);
             },
             intake,
-            lift
+            dispenser
         );
     }
 }
