+++ src/main/java/frc/robot/commands/lift/DispenseIn.java
@@ -2,19 +2,24 @@ package frc.robot.commands.lift;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.constants.DispenserConstants;
-import frc.robot.subsystems.lift.Lift;
+import frc.robot.subsystems.dispenser.Dispenser;
+import frc.robot.subsystems.elevator.Elevator;
 
 public class DispenseIn extends Command {
-    final Lift lift;
+    final Dispenser dispenser;
+    final Elevator elevator;
+
     boolean wasBroken = false;
 
     /**
      * Run the dispenser using intake
      *
-     * @param lift the lift subsystem
+     * @param lift     the lift subsystem
+     * @param elevator the elevator subsystem, not a command requirement
      */
-    public DispenseIn(Lift lift) {
-        this.lift = lift;
+    public DispenseIn(Dispenser lift, Elevator elevator) {
+        this.dispenser = lift;
+        this.elevator = elevator;
 
         addRequirements(lift);
     }
@@ -22,21 +27,21 @@ public class DispenseIn extends Command {
     @Override
     public void initialize() {
         wasBroken = false;
-        lift.dispense(DispenserConstants.Power.kDispenseIn);
+        dispenser.dispense(DispenserConstants.Power.kDispenseIn);
     }
 
     @Override
     public void execute() {
-        wasBroken |= lift.getDispenser().broken;
+        wasBroken |= dispenser.getDispenser().broken;
     }
 
     @Override
     public void end(boolean interrupted) {
-        lift.dispense(0);
+        dispenser.dispense(0);
     }
 
     @Override
     public boolean isFinished() {
-        return !lift.atLevel(0) || (wasBroken && !lift.getDispenser().broken);
+        return !elevator.atLevel(0) || (wasBroken && !dispenser.getDispenser().broken);
     }
 }
