+++ src/main/java/frc/robot/commands/lift/Elevate.java
@@ -4,10 +4,11 @@ import edu.wpi.first.wpilibj2.command.CommandScheduler;
 import frc.robot.commands.util.FunctionWrapper;
 import frc.robot.constants.ElevatorConstants;
 import frc.robot.event.Emitter;
-import frc.robot.subsystems.lift.Lift;
+import frc.robot.subsystems.dispenser.Dispenser;
+import frc.robot.subsystems.elevator.Elevator;
 
 public class Elevate extends FunctionWrapper {
-    private static Emitter.Key<Double, Lift.ReefRange> retract;
+    private static Emitter.Key<Double, Dispenser.ReefRange> retract;
 
     /**
      * Elevate to a certain height. This command will block until the lift is at the desired height.
@@ -15,31 +16,31 @@ public class Elevate extends FunctionWrapper {
      * @param lift   The lift subsystem
      * @param height The height to elevate to
      */
-    public Elevate(Lift lift, int height) {
+    public Elevate(Elevator lift, int height) {
         this(lift, height, true);
     }
 
     /**
      * Elevate to a certain height. Decides whether to block or not.
      *
-     * @param lift     The lift subsystem
+     * @param elevator The elevator subsystem
      * @param height   The height to elevate to
-     * @param blocking Whether to block until the lift is at the desired height
+     * @param blocking Whether to block until the elevator is at the desired height
      */
-    public Elevate(Lift lift, int height, boolean blocking) {
-        super(() -> lift.elevate(height), () -> lift.atLevel(height) || !blocking, lift);
+    public Elevate(Elevator elevator, int height, boolean blocking) {
+        super(() -> elevator.elevate(height), () -> elevator.atLevel(height) || !blocking, elevator);
     }
 
-    public static void registerRetract(Lift lift) {
-        final Elevate instance = new Elevate(lift, 0, true);
+    public static void registerRetract(Elevator elevator) {
+        final Elevate instance = new Elevate(elevator, 0, true);
         final CommandScheduler scheduler = CommandScheduler.getInstance();
 
         retract = Emitter.on(
-            new Lift.ReefRange(),
+            new Dispenser.ReefRange(),
             distance -> {
                 boolean outside = distance >= ElevatorConstants.Range.kRetract;
                 boolean scheduled = scheduler.isScheduled(instance);
-                boolean zeroed = lift.atLevel(0);
+                boolean zeroed = elevator.atLevel(0);
 
                 if (outside && !scheduled && !zeroed) scheduler.schedule(instance);
             }
