+++ src/main/java/frc/robot/commands/lift/ManualElevate.java
@@ -1,30 +1,30 @@
 package frc.robot.commands.lift;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.lift.Lift;
+import frc.robot.subsystems.elevator.Elevator;
 
 public class ManualElevate extends Command {
-    final Lift lift;
+    final Elevator elevator;
     final boolean up;
 
     /**
      * Manually move elevator. - for up, + for down
      *
-     * @param lift the lift subsystem
-     * @param up   whether to move up
+     * @param elevator the elevator subsystem
+     * @param up       whether to move up
      */
-    public ManualElevate(Lift lift, boolean up) {
-        this.lift = lift;
+    public ManualElevate(Elevator elevator, boolean up) {
+        this.elevator = elevator;
         this.up = up;
     }
 
     @Override
     public void initialize() {
-        lift.manualElevate(up);
+        elevator.manualElevate(up);
     }
 
     @Override
     public void end(boolean interrupted) {
-        lift.hold(0.1 * (up ? 1 : -1));
+        elevator.hold(0.1 * (up ? 1 : -1));
     }
 }
