+++ src/main/java/frc/robot/commands/lift/SlowBack.java
@@ -3,34 +3,37 @@ package frc.robot.commands.lift;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.constants.DispenserConstants;
-import frc.robot.subsystems.lift.Lift;
+import frc.robot.subsystems.dispenser.Dispenser;
 
 public class SlowBack extends Command {
-    final Lift lift;
+    final Dispenser dispenser;
     final Timer timer = new Timer(); // todo: replace with encoder readings
 
-    public SlowBack(Lift lift) {
-        this.lift = lift;
-
-        addRequirements(lift);
+    /**
+     * Run the coral back through the dispenser slowly
+     *
+     * @param dispenser the dispenser subsystem
+     */
+    public SlowBack(Dispenser dispenser) {
+        this.dispenser = dispenser;
+        addRequirements(dispenser);
     }
 
     @Override
     public void initialize() {
         timer.stop();
         timer.reset();
-
-        lift.dispense(DispenserConstants.Power.kSlowBack);
+        dispenser.dispense(DispenserConstants.Power.kSlowBack);
     }
 
     @Override
     public void execute() {
-        if (lift.getDispenser().broken) timer.restart();
+        if (dispenser.getDispenser().broken) timer.restart();
     }
 
     @Override
     public void end(boolean interrupted) {
-        lift.dispense(0);
+        dispenser.dispense(0);
     }
 
     @Override
