+++ src/main/java/frc/robot/subsystems/drivetrain/Drivetrain.java
@@ -20,9 +20,9 @@ import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
 import frc.robot.constants.*;
 import frc.robot.event.Emitter;
 import frc.robot.util.IterUtil;
-import frc.robotio.drivetrain.GyroIO;
-import frc.robotio.drivetrain.SwerveIO;
-import frc.robotio.photon.CameraIO;
+import frc.robotio.GyroIO;
+import frc.robotio.SwerveIO;
+import frc.robotio.CameraIO;
 import org.dyn4j.geometry.Vector2;
 import org.json.simple.parser.ParseException;
 import org.littletonrobotics.junction.AutoLogOutput;
