+++ src/main/java/frc/robot/subsystems/intake/Intake.java
@@ -7,7 +7,7 @@ import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.constants.VisionConstants;
 import frc.robot.event.Emitter;
 import frc.robot.subsystems.drivetrain.Drivetrain;
-import frc.robotio.intake.RollerIO;
+import frc.robotio.RollerIO;
 import org.littletonrobotics.junction.Logger;
 
 import java.util.List;
