+++ src/main/java/frc/robotio/CameraIO.java
@@ -1,4 +1,4 @@
-package frc.robotio.photon;
+package frc.robotio;
 
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Pose3d;
