+++ src/main/java/frc/robotio/GyroIO.java
@@ -1,4 +1,4 @@
-package frc.robotio.drivetrain;
+package frc.robotio;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.units.measure.Angle;
