+++ src/main/java/frc/robotreplay/GyroReplay.java
@@ -1,0 +1,11 @@
+package frc.robotreplay;
+
+import frc.robotio.GyroIO;
+
+public class GyroReplay extends GyroIO {
+    @Override
+    public void reset() { }
+
+    @Override
+    public void update() { }
+}
