+++ src/main/java/frc/robotreplay/ModuleReplay.java
@@ -1,0 +1,13 @@
+package frc.robotreplay;
+
+import edu.wpi.first.math.kinematics.SwerveModuleState;
+import frc.robotio.SwerveIO;
+
+public class ModuleReplay extends SwerveIO {
+    /** Keep constructors the same so I can use fn pointers without java screaming */
+    public ModuleReplay(int _id) { }
+
+    public void update() { }
+
+    public void setDesiredState(SwerveModuleState desired) { }
+}
