+++ src/main/java/frc/robotreplay/RollerReplay.java
@@ -1,0 +1,11 @@
+package frc.robotreplay;
+
+import frc.robotio.RollerIO;
+
+public class RollerReplay extends RollerIO {
+	@Override
+	public void set(double left, double right) {}
+
+	@Override
+	public void update() {}
+}
