--- src/main/java/frc/robotreplay/drivetrain/GyroReplay.java
@@ -1,11 +1,0 @@
-package frc.robotreplay.drivetrain;
-
-import frc.robotio.drivetrain.GyroIO;
-
-public class GyroReplay extends GyroIO {
-	@Override
-	public void reset() {}
-
-	@Override
-	public void update() {}
-}
