--- src/main/java/frc/robotreplay/drivetrain/ModuleReplay.java
@@ -1,13 +1,0 @@
-package frc.robotreplay.drivetrain;
-
-import edu.wpi.first.math.kinematics.SwerveModuleState;
-import frc.robotio.drivetrain.SwerveIO;
-
-public class ModuleReplay extends SwerveIO {
-	/** Keep constructors the same so I can use fn pointers without java screaming */
-	public ModuleReplay(int _id) {}
-
-	public void update() {}
-
-	public void setDesiredState(SwerveModuleState desired) {}
-}
