--- src/main/java/frc/robotreplay/intake/RollerReplay.java
@@ -1,11 +1,0 @@
-package frc.robotreplay.intake;
-
-import frc.robotio.intake.RollerIO;
-
-public class RollerReplay extends RollerIO {
-    @Override
-    public void set(double left, double right) { }
-
-    @Override
-    public void update() { }
-}
